Rename new DataStringConverter
