+++ data/src/main/java/com/aconno/sensorics/data/converter/DataStringConverter.kt
@@ -1,71 +1,146 @@
 package com.aconno.sensorics.data.converter
 
 import com.aconno.sensorics.domain.model.Reading
-import java.util.*
 import java.util.regex.Pattern
 
-class DataStringConverter(userDataString: String) {
+class DataStringConverter(
+    private var userDataString: String = ""
+) {
 
     companion object {
         private const val VALUE = "value"
+        private const val NAME = "name"
+        private const val TS = "ts"
+        private const val DEVICE = "device"
+
+        private const val NOT_VALID = -1
+        private const val ONE_BY_ONE = 1
+        private const val CHUNK = 2
     }
 
-    private var map: HashMap<String, String>? = null
-
+    private val list: MutableList<String> = mutableListOf()
     private val pattern = Pattern.compile("\\$\\s*(\\w+)")
+    private var type: Int = 0
 
     init {
-        parseDataString(userDataString)
+        if (!userDataString.isEmpty()) {
+            parseAndValidateDataString(userDataString)
+        }
     }
 
-    private fun parseDataString(everything: String) {
-        map = HashMap()
+    fun parseAndValidateDataString(everything: String): Boolean {
+        val matcher = pattern.matcher(everything)
+        while (matcher.find()) {
+            list.add(matcher.group().replace("$", "").toLowerCase())
+        }
 
-        val scanner = Scanner(everything)
-        while (scanner.hasNextLine()) {
-            val line = scanner.nextLine()
+        type = checkDataStringValid()
+        println("Validation = ${type != NOT_VALID}")
+        return type != NOT_VALID
+    }
 
-            val matcher = pattern.matcher(line)
-            while (matcher.find()) {
-                if (!matcher.group(1).equals(
-                        "ts",
-                        ignoreCase = true
-                    ) && !matcher.group(1).equals("name", ignoreCase = true)
-                ) {
-                    map!![matcher.group(1)] = line
+    private fun checkDataStringValid(): Int {
+        val filteredList = list.filter { it != TS && it != DEVICE } as MutableList<String>
+
+        return if (filteredList.contains(VALUE) || filteredList.contains(NAME)) {
+            filteredList.removeAll { it == VALUE || it == NAME }
+
+            if (filteredList.size == 0) {
+                ONE_BY_ONE
+            } else {
+                NOT_VALID
+            }
+        } else {
+            CHUNK
+        }
+    }
+
+    fun convert(readings: List<Reading>): List<String> {
+        if (type == ONE_BY_ONE) {
+            return getOneByOne(readings, userDataString)
+        } else if (type == CHUNK) {
+            return listOf(getChunk(readings, userDataString))
+        }
+
+        return listOf(userDataString)
+    }
+
+    fun convert(readings: List<Reading>, userDataString: String): List<String> {
+        this@DataStringConverter.userDataString = userDataString
+        if (parseAndValidateDataString(userDataString)) {
+            return convert(readings)
+        }
+
+        return listOf(userDataString)
+    }
+
+    private fun getOneByOne(readings: List<Reading>, userDataString: String): List<String> {
+        val responseList = mutableListOf<String>()
+
+        readings.forEach { reading ->
+            var newDataString = userDataString
+
+            list.forEach {
+                newDataString = when (it) {
+                    NAME -> {
+                        newDataString.replace("\$$it", reading.name)
+                    }
+                    VALUE -> {
+                        newDataString.replace("\$$it", reading.value.toString())
+                    }
+                    TS -> {
+                        newDataString.replace("\$$it", reading.timestamp.toString())
+                    }
+                    DEVICE -> {
+                        newDataString.replace("\$$it", reading.device.macAddress)
+                    }
+                    else -> {
+                        throw IllegalArgumentException()
+                    }
                 }
             }
+
+            responseList.add(newDataString)
         }
-        scanner.close()
+
+        return responseList
     }
 
-    fun convert(data: Reading): List<String>? {
-        val type = data.name.toLowerCase()
-        val map = this.map!!
+    fun convertDeviceAndTS(reading: Reading, userDataString: String): String {
+        return userDataString.replace("\$$TS", reading.timestamp.toString())
+            .replace("\$$DEVICE", reading.device.macAddress)
+    }
 
-        val dataString: String
-
-        dataString = if (map.containsKey(type)) {
-            map[type]!!.replace(
-                "$" + data.name.toLowerCase(),
-                data.value.toString()
-            )
-        } else {
-            if (map.containsKey(DataStringConverter.VALUE)) {
-                map[DataStringConverter.VALUE]!!.replace(
-                    "\$value",
-                    data.value.toString()
-                )
-            } else {
-                return null
+    private fun getChunk(
+        readings: List<Reading>,
+        userDataString: String
+    ): String {
+        var newDataString = userDataString
+        list.forEach {
+            when (it) {
+                TS -> {
+                    newDataString = newDataString.replace("\$$TS", readings[0].timestamp.toString())
+                }
+                DEVICE -> {
+                    newDataString =
+                            newDataString.replace("\$$DEVICE", readings[0].device.macAddress)
+                }
+                else -> {
+                    val find =
+                        readings.find { reading ->
+                            reading.name.replace(
+                                " ",
+                                "_"
+                            ).toLowerCase() == it
+                        }
+                    if (find != null) {
+                        newDataString = newDataString.replace("$$it", find.value.toString())
+                    }
+                }
             }
+
         }
 
-        return listOf(
-            dataString
-                .replace("\$ts", System.currentTimeMillis().toString())
-                .replace("\$name", data.name)
-                .replace("\$device", data.device.macAddress)
-        )
+        return newDataString
     }
 }
